# feat: increase file scanning performance